refactor code into pages
